refactor: consolidate access_keys + precompiles into keychain module
Per Oracle review, this reorganization:
- Renames precompiles.py → keychain.py (clearer domain naming)
- Merges access_keys.py into keychain.py (keychain-specific signing)
- Removes format_spending_limit() (presentation concern, belongs in consumer apps)

The keychain module now contains all AccountKeychain precompile functionality:
- Precompile address and function selectors
- Calldata encoding for getRemainingLimit()
- Spending limit queries
- Keychain signature constants (0x03 type, 86-byte format)
- build_keychain_signature() and sign_tx_access_key()

All exports remain available from pytempo root for backwards compatibility.
